Split symmetric support into separate files
re #6 re #7
